consolidate CGSize extensions
