--- BlueprintUI/Sources/Internal/Extensions/CGSize+Arithmetic.swift
@@ -1,11 +1,0 @@
-import CoreGraphics
-
-extension CGSize {
-    static func + (lhs: CGSize, rhs: CGSize) -> CGSize {
-        CGSize(width: lhs.width + rhs.width, height: lhs.height + rhs.height)
-    }
-
-    static func - (lhs: CGSize, rhs: CGSize) -> CGSize {
-        CGSize(width: lhs.width - rhs.width, height: lhs.height - rhs.height)
-    }
-}
