+++ BlueprintUI/Sources/Layout/Inset.swift
@@ -180,12 +180,3 @@ extension Inset {
         }
     }
 }
-
-extension CGSize {
-    func inset(by insets: UIEdgeInsets) -> Self {
-        CGSize(
-            width: width - insets.left - insets.right,
-            height: height - insets.top - insets.bottom
-        )
-    }
-}
